requirements.txt
streamlit==1.37.0
transformers==4.41.2
torch==2.3.0
torchvision==0.18.0
pillow==10.4.0
wikipedia==1.4.0
requests==2.32.3
beautifulsoup4==4.12.3
matplotlib==3.9.2
numpy==1.26.4
